server-api/am/requirements.txt
flask-restplus
Flask-SQLAlchemy==2.1
flask_cors
